+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gambit",
+  "name": "tybalt",
   "version": "0.0.1",
   "description": "",
   "private": true,
