+++ packages/cli/src/commands/scaffold.ts
@@ -37,9 +37,9 @@ const action = (componentName: string, options: ScaffoldCommandOptions) => {
 
 export default ({ program }: CommandContext) => {
     program.command('scaffold')
-        .description('scaffold gambit-related files')
+        .description('scaffold tybalt-related files')
         .argument('<string>', 'target', 'component')
-        .option('-n, --name', 'gambit-example')
+        .option('-n, --name', 'tybalt-example')
         .option('-s, --styles', 'whether to generate a css module', true)
         .option('--template', 'whether to generate an html template', true)
         .option('-t, --tests', 'whether to generate unit tests', true)
