+++ packages/eleventy-plugin/index.js
@@ -1,1 +1,1 @@
-console.log("gambit eleventy plugin loaded");
+console.log("tybalt eleventy plugin loaded");
