+++ packages/example/components/CookieBanner/cookie-banner.examples.ts
@@ -1,4 +1,4 @@
-import { defineExample } from '@gambit/core';
+import { defineExample } from '@tybalt/core';
 import CookieBanner from './cookie-banner.component';
 
 export default {
