+++ packages/test-utils-lit-example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gambit/test-utils-lit-example",
+  "name": "@tybalt/test-utils-lit-example",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -13,7 +13,7 @@
   "author": "Douglas Wade <douglas.b.wade@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@gambit/test-utils": "^0.0.1",
+    "@tybalt/test-utils": "^0.0.1",
     "@types/jest": "^29.2.3",
     "jest": "^29.2.2",
     "jest-environment-jsdom": "^29.2.2",
