+++ packages/validator/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@gambit/validator",
+    "name": "@tybalt/validator",
     "version": "0.0.1",
     "description": "A library for performing asynchronous validations",
     "main": "dist/index.js",
