+++ packages/website/.eleventy.js
@@ -1,12 +1,12 @@
-// const gambitPlugin = require("@gambit/eleventy-plugin");
+// const tybaltPlugin = require("@tybalt/eleventy-plugin");
 
 module.exports = function (eleventyConfig) {
   eleventyConfig.addPassthroughCopy("./components");
   eleventyConfig.addPassthroughCopy("./lib");
-  //   eleventyConfig.addPlugin(gambitPlugin, {
+  //   eleventyConfig.addPlugin(tybaltPlugin, {
   //     componentModules: ["./components/index.js"],
   //   });
   return {
-    pathPrefix: "/gambit/",
+    pathPrefix: "/tybalt/",
   };
 };
