+++ packages/website/components/code-block.js
@@ -1,6 +1,6 @@
-import { defineComponent } from "../lib/gambit-core/index.js";
+import { defineComponent } from "../lib/tybalt-core/index.js";
 
 defineComponent({
-  name: "gambit-code-block",
+  name: "tybalt-code-block",
   template: "<code>code block</code>",
 });
