+++ packages/website/index.html
@@ -1,12 +1,12 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Gambit</title>
+    <title>Tybalt</title>
   </head>
   <body>
-    <gambit-header></gambit-header>
-    <gambit-code-block></gambit-code-block>
-    <gambit-footer></gambit-footer>
+    <tybalt-header></tybalt-header>
+    <tybalt-code-block></tybalt-code-block>
+    <tybalt-footer></tybalt-footer>
   </body>
   <script type="module" src="{{ '/components/index.js' | url }}"></script>
 </html>
