+++ packages/website/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@gambit/website",
+  "name": "@tybalt/website",
   "version": "0.0.1",
-  "description": "The gambit documentation website",
+  "description": "The tybalt documentation website",
   "main": "index.js",
   "private": true,
   "scripts": {
-    "build": "mkdir -p lib/ && cp -r ../core/dist lib/gambit-core && eleventy",
+    "build": "mkdir -p lib/ && cp -r ../core/dist lib/tybalt-core && eleventy",
     "clean": "rimraf lib/ _site/",
     "serve": "eleventy --serve --port 8081",
     "test": "echo \"Error: no test specified\" && exit 0"
@@ -18,7 +18,7 @@
     "rimraf": "^3.0.2"
   },
   "dependencies": {
-    "@gambit/core": "^0.0.1",
-    "@gambit/eleventy-plugin": "^0.0.1"
+    "@tybalt/core": "^0.0.1",
+    "@tybalt/eleventy-plugin": "^0.0.1"
   }
 }
