+++ packages/website/pages/core.md
@@ -1,1 +1,1 @@
-# @gambit/core
+# @tybalt/core
