+++ packages/website/pages/test-utils.md
@@ -1,1 +1,1 @@
-# @gambit/test-utils
+# @tybalt/test-utils
