+++ packages/website/pages/validator.md
@@ -1,1 +1,1 @@
-# @gambit/validator
+# @tybalt/validator
